+++ packages/pyroscope-panel-plugin/.eslintrc.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  extends: [path.join(__dirname, '../../.eslintrc.js')],
+  ignorePatterns: ['.eslintrc.js'],
+};