+++ webapp/javascript/components/Sidebar.tsx
@@ -221,7 +221,7 @@ export function SidebarComponent() {
             </MenuItem>
           )}
           <MenuItem icon={<Icon icon={faInfoCircle} />}>
-            Status
+            Scrape Targets
             <NavLink to={{ pathname: PAGES.SERVICE_DISCOVERY, search }} exact />
           </MenuItem>
           <MenuItem icon={<Icon icon={faFileAlt} />}>